fetch rating components only for requested players
`player_ratings_components_for_release` was building a lookup hash for all players in a release, which used a lot of memory. We now filter by the list of players present on a JSON response page.

We build a list of placeholders for `in ($2, $3, $4)` in two places now, so this got moved into a `build_placeholders` method.